refactor page content into separate module
